Yidun/requirements.txt
cryptography==37.0.3
Flask==2.1.2
loguru==0.5.3
opencv_python_headless==4.5.3.56
Pillow==9.3.0
pycryptodome==3.16.0
PyExecJS==1.5.1
redis==4.1.3
requests==2.27.1
chardet==3.0.4